+++ Task_2_1_1/src/main/java/org/example/hasnotprimecheck/ChartBuilder.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.hasnotprimecheck;
 
 import javax.swing.*;
 import java.awt.*;
@@ -27,7 +27,7 @@ public class ChartBuilder extends JFrame {
     private void initUi() {
         setTitle("Execution Time Chart");
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setSize(800, 600);
+        setSize(1000, 600);
 
         ChartPanel chartPanel = createChartPanel();
         add(chartPanel, BorderLayout.CENTER);
