+++ Task_2_1_1/src/main/java/org/example/hasnotprimecheck/checkers/ParallelWithStreamChecker.java
@@ -1,0 +1,20 @@
+package org.example.hasnotprimecheck.checkers;
+
+import java.util.Arrays;
+
+public class ParallelWithStreamChecker extends Checker {
+    /**
+     * hasNonPrime.
+     */
+    public static boolean hasNonPrime(int[] nums) {
+        return Arrays.stream(nums).parallel().anyMatch(num -> !isPrime(num));
+    }
+
+    /**
+     * hasNonPrime.
+     */
+    public static boolean hasNonPrime(int[] nums, int numThreads) {
+        System.setProperty("java.util.concurrent.ForkJoinPool.common.parallelism", String.valueOf(numThreads));
+        return Arrays.stream(nums).parallel().anyMatch(num -> !isPrime(num));
+    }
+}
