+++ Task_2_1_1/src/main/java/org/example/hasnotprimecheck/checkers/SequentialChecker.java
@@ -1,0 +1,12 @@
+package org.example.hasnotprimecheck.checkers;
+
+import java.util.Arrays;
+
+public class SequentialChecker extends Checker {
+    /**
+     * hasNonPrime.
+     */
+    public static boolean hasNonPrime(int[] nums) {
+        return Arrays.stream(nums).sequential().anyMatch(num -> !isPrime(num));
+    }
+}
