+++ Task_2_1_1/src/test/java/org/example/hasnotprimecheck/checkers/CheckerForThreadTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.hasnotprimecheck.checkers;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
